# swagger.yaml
swagger: "2.0"
info:
  description: "Datafuel API docs"
  version: "1.0.0"
  title: "Datafuel"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "hellokalash@datafuel.ru"
host: "api.datafuel.ru"
basePath: "/"
...
schemes:
- "https"
securityDefinitions:
  basicAuth:
    type: basic
security:
  - basicAuth: []
paths:
...
  /stats/users:
    post:
      tags:
      - "VK stats"
      summary: "Get users stats by list of IDs"
      description: "Returns a stats for multiple users"
      produces:
      - "application/json"
      consumes:
      - "application/json"
      parameters:
      - name: "body"
        in: "body"
        description: "Request parameters"
        schema:
          type: "object"
          required: 
            - "ids"
          properties:
            ids:
              description: "Users' IDs to return"
              type: "array"
              items:
                type: "integer"
                format: "int64"
                minimum: 1
            filters:
              description: "Filters for basic users selecting"
              type: "object"
              properties:
                age_bounds:
                  type: "array"
                  items:
                    type: "object"
                    properties:
                      from:
                        type: "integer"
                        format: "int8"
                        minimum: 1
                        maximum: 100
                        default: 1
                      to:
                        type: "integer"
                        format: "int8"
                        minimum: 1
                        maximum: 100
                        default: 100
                city_ids:
                  type: "array"
                  items:
                    type: "integer"
                    format: "int64"
                country_ids:
                  type: "array"
                  items:
                    type: "integer"
                    format: "int64"
                sex:
                  type: "array"
                  items:
                    type: "string"
                    enum: ["male", "female"]
                    format: "str"
                MBTIs:
                  type: "array"
                  items:
                    type: "string"
                    enum: ["ISTJ", "ISTP", "ISFJ", "ISFP", "INTJ", "INTP", "INFJ", "INFP", "ESTJ", "ESTP", "ESFJ", "ESFP", "ENTJ", "ENTP", "ENFJ", "ENFP"]
                    format: "str"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Stats"
        400:
          description: "Error in requests params"
          schema:
            $ref: "#/definitions/Error"
        402:
          description: "No more requests left"
        404:
          description: "User not found"
          schema:
            $ref: "#/definitions/Error"
...
definitions:
...
  Stats:
    type: "object"
    properties:
      age:
        type: "object"
        properties:
          mean:
            type: "number"
            format: "double"
          std:
            type: "number"
            format: "double"
          not_null_count:
            type: "integer"
            format: "int64"
      friends_count:
        type: "object"
        properties:
          mean:
            type: "number"
            format: "double"
          std:
            type: "number"
            format: "double"
          not_null_count:
            type: "integer"
            format: "int64"
      groups_count:
        type: "object"
        properties:
          mean:
            type: "number"
            format: "double"
          std:
            type: "number"
            format: "double"
          not_null_count:
            type: "integer"
            format: "int64"
      topn_groups:
        type: "array"
        items:
            type: "object"
            properties:
              id:
                type: "integer"
                format: "int64"
              count:
                type: "integer"
                format: "int64"
      MBTI:
        type: "object"
        properties:
          types_distribution:
            type: "object"
            properties:
              ISTJ:
                type: "integer"
                format: "int64"
              ISTP:
                type: "integer"
                format: "int64"
              ISFJ:
                type: "integer"
                format: "int64"
              ISFP:
                type: "integer"
                format: "int64"
              INTJ:
                type: "integer"
                format: "int64"
              INTP:
                type: "integer"
                format: "int64"
              INFJ:
                type: "integer"
                format: "int64"
              INFP:
                type: "integer"
                format: "int64"
              ESTJ:
                type: "integer"
                format: "int64"
              ESTP:
                type: "integer"
                format: "int64"
              ESFJ:
                type: "integer"
                format: "int64"
              ESFP:
                type: "integer"
                format: "int64"
              ENTJ:
                type: "integer"
                format: "int64"
              ENTP:
                type: "integer"
                format: "int64"
              ENFJ:
                type: "integer"
                format: "int64"
              ENFP:
                type: "integer"
                format: "int64"
          basis_distribution:
            type: "object"
            properties:
              IE:
                type: "object"
                properties:
                  mean:
                    type: "number"
                    format: "double"
              SN:
                type: "object"
                properties:
                  mean:
                    type: "number"
                    format: "double"
              TF:
                type: "object"
                properties:
                  mean:
                    type: "number"
                    format: "double"
              JP:
                type: "object"
                properties:
                  mean:
                    type: "number"
                    format: "double"
  Error:
    type: "object"
    properties:
      error:
        type: "string"
        format: "str"